+++ temba/api/v2/views.py
@@ -21,9 +21,10 @@ from temba.flows.models import Flow, FlowRun, FlowStart, FlowStartCount
 from temba.globals.models import Global
 from temba.locations.models import AdminBoundary, BoundaryAlias
 from temba.msgs.models import Broadcast, BroadcastMsgCount, Label, LabelCount, Media, Msg, MsgFolder, OptIn
-from temba.orgs.models import OrgMembership, User
+from temba.orgs.models import OrgMembership
 from temba.orgs.views.mixins import OrgPermsMixin
 from temba.tickets.models import Ticket, Topic
+from temba.users.models import User
 from temba.utils import str_to_bool
 from temba.utils.db.queries import SubqueryCount, or_list
 from temba.utils.uuid import is_uuid
@@ -3092,7 +3093,7 @@ class FlowStartsEndpoint(ListAPIMixin, WriteAPIMixin, BaseEndpoint):
      * **urns** - the URNs you want to start in this flow (array of up to 100 strings, optional)
      * **restart_participants** - whether to restart participants already in this flow (optional, defaults to true)
      * **exclude_active** - whether to exclude contacts currently in other flow (optional, defaults to false)
-     * **params** - extra parameters to pass to the flow start (object, accessible via `@trigger.params` in the flow)
+     * **params** - extra parameters to pass to the flow start (object, must be at most 10K characters, accessible via `@trigger.params` in the flow)
 
     Example: