+++ temba/settings_common.py
@@ -245,11 +245,10 @@ INSTALLED_APPS = (
     "rest_framework.authtoken",
     "compressor",
     "smartmin",
-    "smartmin.users",
     "timezone_field",
+    "temba.users",
     "temba.apks",
     "temba.archives",
-    "temba.auth_tweaks",
     "temba.api",
     "temba.request_logs",
     "temba.classifiers",
@@ -477,8 +476,6 @@ GROUP_PERMISSIONS = {
         "orgs.org_update",
         "orgs.org_workspace",
         "orgs.orgimport.*",
-        "orgs.user_list",
-        "orgs.user_update",
         "request_logs.httplog_list",
         "request_logs.httplog_read",
         "request_logs.httplog_webhooks",
@@ -488,6 +485,8 @@ GROUP_PERMISSIONS = {
         "tickets.ticket.*",
         "tickets.topic.*",
         "triggers.trigger.*",
+        "users.user_list",
+        "users.user_update",
     ),
     "Editors": (
         "airtime.airtimetransfer_list",
@@ -590,7 +589,7 @@ GROUP_PERMISSIONS = {
 
 # extra permissions that only apply to API requests (wildcard notation not supported here)
 API_PERMISSIONS = {
-    "Editors": ("orgs.org_list", "orgs.user_list"),
+    "Editors": ("orgs.org_list", "users.user_list"),
     "Agents": (
         "contacts.contact_create",
         "contacts.contact_list",
@@ -602,8 +601,8 @@ API_PERMISSIONS = {
         "msgs.msg_create",
         "orgs.org_list",
         "orgs.org_read",
-        "orgs.user_list",
         "tickets.shortcut_list",
+        "users.user_list",
     ),
 }
 
@@ -616,6 +615,7 @@ LOGIN_REDIRECT_URL = "/org/choose/"
 
 AUTHENTICATION_BACKENDS = ("temba.orgs.backend.AuthenticationBackend",)
 
+AUTH_USER_MODEL = "users.User"
 AUTH_PASSWORD_VALIDATORS = [
     {"NAME": "django.contrib.auth.password_validation.MinimumLengthValidator", "OPTIONS": {"min_length": 8}},
 ]
@@ -848,8 +848,6 @@ MAILROOM_AUTH_TOKEN = None
 # Data Model
 # -----------------------------------------------------------------------------------
 
-MSG_FIELD_SIZE = 640  # used for broadcast text, message text, and message campaign events
-FLOW_START_PARAMS_SIZE = 256  # used for params passed to flow start API endpoint
 GLOBAL_VALUE_SIZE = 10_000  # max length of global values
 
 ORG_LIMIT_DEFAULTS = {